FT: Componentizando a tela de Todo
